folder refactor, bot folder
