+++ .gitignore
@@ -1,7 +1,6 @@
 .vscode/
 node_modules/
 package-lock.json
-temp.js
 scratch.js
-agent_code/**
 !agent_code/template.js
+bots/**/action-code/**
