+++ bots/andy/memories.json
@@ -1,0 +1,80 @@
+{
+    "name": "andy",
+    "bio": "You are playing minecraft and assisting other players in tasks.",
+    "memory": "",
+    "events": [],
+    "turns": [
+        {
+            "role": "user",
+            "content": "sumixam0: come here andy"
+        },
+        {
+            "role": "assistant",
+            "content": "```\n// I am going to navigate to sumixam0. \nawait skills.goToPlayer(bot, 'sumixam0');\n```"
+        },
+        {
+            "role": "system",
+            "content": "Code output:\nYou have reached at 300.7930424155586, 78, -115.19151414785698.\nPlayer is now at (300.7930424155586, 78, -115.19151414785698).\nCode finished.\n"
+        },
+        {
+            "role": "assistant",
+            "content": "I'm here, sumixam0! What can I do for you?"
+        },
+        {
+            "role": "user",
+            "content": "sumixam0: follow me!"
+        },
+        {
+            "role": "assistant",
+            "content": "Sure, I'll follow you, Sumixam0!\n```\n// I am going to follow sumixam0.\nawait skills.followPlayer(bot, 'sumixam0');\n```"
+        },
+        {
+            "role": "user",
+            "content": "sumixam0: place some dirt"
+        },
+        {
+            "role": "assistant",
+            "content": "Sure, I'll place some dirt.\n```\n// I am going to place dirt at my current position.\nlet position = world.getPosition(bot);\nawait skills.placeBlock(bot, 'dirt', position.x, position.y, position.z);\n```"
+        },
+        {
+            "role": "system",
+            "content": "Code output:\nDon't have any dirt to place.\nCode finished.\n"
+        },
+        {
+            "role": "assistant",
+            "content": "I apologize, but I don't have any dirt to place at the moment."
+        },
+        {
+            "role": "user",
+            "content": "sumixam0: try again"
+        },
+        {
+            "role": "assistant",
+            "content": "Sure, let me try again.\n```\n// I am going to place dirt at my current position.\nlet position = world.getPosition(bot);\nawait skills.placeBlock(bot, 'dirt', position.x, position.y - 1, position.z);\n```"
+        },
+        {
+            "role": "system",
+            "content": "Code output:\nCannot place block at (303, 77, -116) because grass_block is in the way.\nCode finished.\n"
+        },
+        {
+            "role": "assistant",
+            "content": "I apologize, but I am unable to place the dirt block here because there is grass in the way."
+        },
+        {
+            "role": "user",
+            "content": "sumixam0: try placing one higher"
+        },
+        {
+            "role": "assistant",
+            "content": "Alright, I'll try placing the dirt block one block higher.\n```\n// I am going to place dirt one block higher than my current position.\nlet position = world.getPosition(bot);\nawait skills.placeBlock(bot, 'dirt', position.x, position.y, position.z);\n```"
+        },
+        {
+            "role": "system",
+            "content": "Code output:\nSuccessfully placed dirt at (303, 78, -116).\nCode finished.\n"
+        },
+        {
+            "role": "assistant",
+            "content": "I have successfully placed a dirt block one block higher."
+        }
+    ]
+}
